Remove opSlice in favor of opApply.
The current implementation of opSlice did not allow assigning values by ref in
a foreach loop, which should be possible.

As Enumap is a pure value type, implementing foreach via opSlice would be risky
(requiring a pointer that could outlive the stack allocated underlying store).

Instead, foreach is now implemented with opApply, which supports setting values
by ref.

opSlice was removed as the meaning is unclear for Enumap -- should it return
values? (key,value) pairs? Instead, slicing should be done via `byValue`,
`byKey`, or `byKeyValue`.

As a side note, there was some weirdness around making opApply friendly to both
@nogc and non-@nogc environments.
This is achieved by unconditionally casting the delegate to @nogc, even though
this shouldn't work (?) if the delegate is not nogc.